pyproject.toml
[project]
name = "voice-memo-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audio-recorder-streamlit>=0.0.10",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.45.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.2",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "streamlit>=1.52.2",
]
